Removed code duplication Implemented all functions of the fbclient to the dynamic version
