+++ src/statement.rs
@@ -43,7 +43,7 @@ impl<'a> Statement<'a> {
                 tr.handle.as_ptr(),
                 &mut handle,
                 sql.len() as u16,
-                sql.as_ptr() as *const i8,
+                sql.as_ptr() as *const _,
                 3, // TODO: Add a way to select the dialect (1, 2 or 3)
                 &mut *xsqlda,
             ) != 0
@@ -160,7 +160,7 @@ impl<'a> Statement<'a> {
                 tr.conn.handle.as_ptr(),
                 tr.handle.as_ptr(),
                 sql.len() as u16,
-                sql.as_ptr() as *const i8,
+                sql.as_ptr() as *const _,
                 3, // Dialect
                 &*params.xsqlda,
             ) != 0
